+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/router/TransitRouter.java
@@ -247,7 +247,6 @@ public class TransitRouter {
     var nearbyStops = AccessEgressRouter.streetSearch(
       accessRequest,
       temporaryVerticesContainer,
-      serverContext.transitService(),
       streetRequest,
       serverContext.dataOverlayContext(accessRequest),
       type.isEgress(),