+++ src/main/java/org/opentripplanner/updater/trip/MqttGtfsRealtimeUpdater.java
@@ -1,5 +1,8 @@
 package org.opentripplanner.updater.trip;
 
+import static org.opentripplanner.updater.trip.UpdateIncrementality.DIFFERENTIAL;
+import static org.opentripplanner.updater.trip.UpdateIncrementality.FULL_DATASET;
+
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.transit.realtime.GtfsRealtime;
 import java.net.URI;
@@ -139,7 +142,7 @@ public class MqttGtfsRealtimeUpdater implements GraphUpdater {
     @Override
     public void messageArrived(String topic, MqttMessage message) {
       List<GtfsRealtime.TripUpdate> updates = null;
-      boolean fullDataset = true;
+      UpdateIncrementality updateIncrementality = FULL_DATASET;
       try {
         // Decode message
         GtfsRealtime.FeedMessage feedMessage = GtfsRealtime.FeedMessage.PARSER.parseFrom(
@@ -156,7 +159,7 @@ public class MqttGtfsRealtimeUpdater implements GraphUpdater {
             .getIncrementality()
             .equals(GtfsRealtime.FeedHeader.Incrementality.DIFFERENTIAL)
         ) {
-          fullDataset = false;
+          updateIncrementality = DIFFERENTIAL;
         }
 
         // Create List of TripUpdates
@@ -177,7 +180,7 @@ public class MqttGtfsRealtimeUpdater implements GraphUpdater {
             snapshotSource,
             fuzzyTripMatcher,
             backwardsDelayPropagationType,
-            fullDataset,
+            updateIncrementality,
             updates,
             feedId,
             recordMetrics